test(api): 修正接口测试边界与分类
