+++ tests/algorithm/transfer/test_wsb.py
@@ -12,7 +12,7 @@ import numpy as np
 import pytest
 
 from e2m2e.algorithm.dynamics import BCR4BP_Dynamics, BCR4BPSystem, CR3BP_Dynamics, CR3BP_System
-from e2m2e.algorithm.transfer import WsbSearchParams, WsbTransferDetails, transfer_orbit
+from e2m2e.algorithm.transfer import WsbSearchParams, transfer_orbit
 from e2m2e.algorithm.transfer.hohmann import TliParams
 from e2m2e.algorithm.transfer.wsb import compute_kepler_energy_moon, search_wsb_trajectories
 from e2m2e.data.constants import Datum
@@ -170,41 +170,6 @@ class TestWsbSearchParams:
         """perilune_alt_min >= perilune_alt_max 报错。"""
         with pytest.raises(ValueError, match="perilune_alt_min"):
             WsbSearchParams(perilune_alt_min=5000.0, perilune_alt_max=5000.0)
-
-
-class TestWsbTransferDetails:
-    """WSB 搜索结果的数据契约。"""
-
-    def test_preserves_result_payload_and_status_triplet(self):
-        details = WsbTransferDetails(
-            tli_epoch="2025-01-01T00:00:00",
-            tof_sec=1e7,
-            perilune_alt_km=100.0,
-            perilune_vel_km_s=2.5,
-            perilune_state=np.zeros(6),
-            h2_kepler=-0.5,
-            dv_departure_km_s=3.1,
-            dv_arrival_km_s=0.8,
-            n_candidates_searched=100,
-            n_candidates_feasible=5,
-            status=ConvergenceState.CONVERGED,
-            cause=FailureCause.NONE,
-            message="收敛",
-            search_params=WsbSearchParams(),
-        )
-
-        assert isinstance(details.tof_sec, float)
-        assert isinstance(details.perilune_alt_km, float)
-        assert isinstance(details.perilune_vel_km_s, float)
-        assert isinstance(details.dv_departure_km_s, float)
-        assert isinstance(details.dv_arrival_km_s, float)
-        assert isinstance(details.h2_kepler, float)
-        assert isinstance(details.n_candidates_searched, int)
-        assert isinstance(details.n_candidates_feasible, int)
-        assert details.status is ConvergenceState.CONVERGED
-        assert details.cause is FailureCause.NONE
-        assert details.message == "收敛"
-        assert isinstance(details.search_params, WsbSearchParams)
 
 
 # ---------------------------------------------------------------------------
